# Conflicts:
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/cashback/domain/SetCashbackUseCase.kt

+++ features/promo/promo_checkout_common/src/main/java/com/tokopedia/promocheckout/common/data/entity/request/Order.kt
@@ -13,7 +13,7 @@ import kotlinx.android.parcel.Parcelize
 data class Order(
         @SuppressLint("Invalid Data Type")
         @SerializedName("shop_id")
-        var shopId: Int? = 0,
+        var shopId: Long? = 0,
 
         @SerializedName("unique_id")
         var uniqueId: String? = null,
@@ -26,13 +26,13 @@ data class Order(
 
         @SuppressLint("Invalid Data Type")
         @SerializedName("shipping_id")
-        var shippingId: Int? = null,
+        var shippingId: Long? = null,
 
         @SuppressLint("Invalid Data Type")
         @SerializedName("sp_id")
-        var spId: Int? = null,
+        var spId: Long? = null,
 
         @SuppressLint("Invalid Data Type")
         @SerializedName("is_insurance_price")
-        var isInsurancePrice: Int? = null
+        var isInsurancePrice: Long? = null
 ) : Parcelable
